refactor(util): add and use getIconNameFromStatus function
