+++ src/components/InlineNotification/InlineNotification-v2.tsx
@@ -1,9 +1,10 @@
 import clsx from 'clsx';
 import React from 'react';
 
+import getIconNameFromStatus from '../../util/getIconNameFromStatus';
 import type { Status } from '../../util/variant-types';
 
-import Icon, { type IconName } from '../Icon';
+import Icon from '../Icon';
 import Text from '../Text';
 
 import styles from './InlineNotification-v2.module.css';
@@ -34,21 +35,6 @@ type InlineNotificationProps = {
    */
   title: string;
 };
-
-/**
- * Map statuses to existing icon names
- * @param status component status
- * @returns the matching icon name
- */
-function getIconNameFromStatus(status: Status): IconName {
-  const map: Record<Status, IconName> = {
-    informational: 'info',
-    critical: 'dangerous',
-    warning: 'warning',
-    favorable: 'check-circle',
-  };
-  return map[status];
-}
 
 /**
  * `import {InlineNotification} from "@chanzuckerberg/eds";`
